Able to generate style, but need to refactor map initialization method
Because we want to use the incomming DOM events before generating the styles, we'll need catch them and pass them with the initial creation.
Maplibre doesn't allow that a empty object is passed along while initialzing the map AND also doesn't allow adding new sources until the style itself is load.
(This probably has to do with the fact that the style can be a remote URL, so that it is async)

I've added a bunch of TODO that we'll need to fix before this will work.